+++ Engine/renderer/headers/circle.hpp
@@ -4,7 +4,7 @@
 
 class Circle : public Shape {
     public:
-        Circle(Vec2<int>& center, int r, Vec2<int>& v, int m) : mCenter(center), mr(r), mv(v), mMass(m) { }
+        Circle(Vec2<int>& center, int r, Vec2<double>& v, int m) : mCenter(center), mr(r), mv(v), mMass(m) { }
         ~Circle() {}
 
         /* Render */
@@ -24,13 +24,13 @@ class Circle : public Shape {
 
         // Getters
         Vec2<int> getCenter() const override;
-        Vec2<int> getVelocity() const override;
+        Vec2<double> getVelocity() const override;
         int getRadius() const override { return mr; }
         int getMass() const override { return mMass; }
 
     private:
         Vec2<int> mCenter;
-        Vec2<int> mv;
+        Vec2<double> mv;
         int mr;
         int mMass;
 };
